requirements.txt
numpy 
panda
matplotlib
seaborn
plotly